requirements.txt
numpy==1.23.5
scipy==1.10.1
opencv-python-headless==4.7.0.72
imgaug==0.4.0
imageio==2.31.1
matplotlib==3.8.4
Pillow==9.5.0
six==1.16.0
PyYAML==6.0.2
tqdm==4.66.4
